+++ .nvim.lua
@@ -1,0 +1,7 @@
+require("flutter-tools").setup_project {
+  {
+    name = "Example",
+    target = "lib/main.dart",
+    cwd = "example",
+  },
+}