+++ ios/Classes/FlutterEmailSenderPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterEmailSenderPlugin : NSObject<FlutterPlugin>
-@end